Refactor container factory into smaller files
The number of methods was confusing.
